Lockscreen charging info: show decimal in battery temperature

+++ packages/SystemUI/src/com/android/systemui/statusbar/KeyguardIndicationController.java
@@ -200,7 +200,7 @@ public class KeyguardIndicationController {
 
     private int mChargingCurrent;
     private double mChargingVoltage;
-    private int mTemperature;
+    private float mTemperature;
 
     private KeyguardUpdateMonitorCallback mUpdateMonitorCallback;
